remove old comments; change gpu indices from int to size_t

+++ device_info.cpp
@@ -3,8 +3,8 @@
 
 namespace {
   // returns a vector of indices of "accelerators" that are actual GPU devices 
-  std::vector<int> get_gpu_indices(const std::vector<hc::accelerator>& accelerators) {
-    std::vector<int> gpu_indices;
+  std::vector<size_t> get_gpu_indices(const std::vector<hc::accelerator>& accelerators) {
+    std::vector<size_t> gpu_indices;
     for(std::size_t i=0; i!= accelerators.size(); ++i){
       if(accelerators[i].get_device_path() != L"cpu"){
 	gpu_indices.push_back(i);
